[menu][Android] Remove unused code
